Moved intentDao to interactor
