+++ back-end/api/controllers/transcript_controller.js
@@ -1,17 +1,7 @@
-import {IntentInterface} from "../../interfaces/intent-interface.js";
 import {InputBoundaryInterface} from "../../interfaces/input-boundary-interface.js";
 
 export default class TranscriptController {
-    static #intentDao;
     static #transcriptFormatter;
-
-    static setIntentDao(dao) {
-        if(dao.isIntentInterface){
-            this.#intentDao = dao;
-        } else {
-            throw new Error("not an IntentInterface");
-        }
-    }
 
     static setTranscriptFormatter(formatter) {
         if(formatter.isInputBoundaryInterface){
@@ -24,7 +14,7 @@ export default class TranscriptController {
     static async getTranscript(req, res, next) {
 
         const query = req.body;
-        const {intentList} = await this.#intentDao.getIntent(query);
+        const {intentList} = await this.#transcriptFormatter.getTranscript(query);
 
         res.json(intentList);
     }
@@ -33,8 +23,7 @@ export default class TranscriptController {
         try {
             const body = JSON.stringify(req.body.payload);
             const transcript = JSON.parse(body);
-            const content = await this.#transcriptFormatter.formatTranscript(transcript);
-            await this.#intentDao.postIntents(content);
+            await this.#transcriptFormatter.formatTranscript(transcript);
             res.status(200).json({ message: "success" })
         } catch (e) {
             res.status(500).json({ error: e.message })
