+++ back-end/interfaces/input-boundary-interface.js
@@ -8,10 +8,19 @@ export class InputBoundaryInterface {
     static isInputBoundaryInterface = true;
 
     /**
+     * @param {Object} query find specific intent(s)
+     */
+    static async getTranscript(query = {}){
+        throw new Error("not implemented");
+    }
+
+    /**
      *  @param {Object} rawTranscript raw transcript uploaded by user
      *  @return {Object} properly formatted data from the transcript
      * */
     async formatTranscript(rawTranscript = {}) {
         throw new Error("not implemented");
     }
+
+
 }
